Cau hinh ruff + don sach lint + cap nhat README
ruff.toml: 137 loi ban dau, trong do 94 la EXE002 duong tinh gia tren bind mount
Windows (moi file mang bit executable nen ruff doi shebang cho ca module thu vien).
Cac ngoai le con lai deu co ly do ghi kem:
- PLC0415 import trong ham: CO CHU Y o adapter, whisperx/demucs/voxcpm/yt_dlp chi
  co trong image worker. Import o dau module thi image api khong khoi dong duoc
- RUF001/002/003: toan bo tai lieu viet bang tieng Viet, dung dau - va " " that
- PLW0603 bien global: cache model la co y, nap VoxCPM2 mat hang chuc giay
- E501 rieng ffmpeg.py: hai dong Format:/Style: cua file ASS BAT BUOC nam tren mot
  dong theo dac ta SubStation Alpha, xuong dong la file phu de hong

Mot bug that: tests/fakes.py dung datetime.now() khong timezone (DTZ005). So sanh
voi expires_at co timezone se nem TypeError - va cho nem ra la trong license gate.

ruff.toml duoc mount vao container (thieu no thi cau hinh khong co tac dung).
Ket qua: ruff check src tests -> All checks passed.

README: them make migrate (bat buoc lan dau), make test-int, link /web, muc Test,
quy tac nghiep vu thu 6 ve xac minh chu so huu.

191 unit + 43 integration test xanh.

+++ src/infrastructure/tts/voxcpm.py
@@ -48,7 +48,7 @@ def _load_model(model_dir: Path | None) -> Any:
     log.info("voxcpm.load.start", model_dir=str(model_dir) if model_dir else "default")
     try:
         _model = VoxCPM.from_pretrained(str(model_dir)) if model_dir else VoxCPM.from_pretrained()
-    except Exception as exc:  # noqa: BLE001 — lỗi nạp model rất đa dạng
+    except Exception as exc:
         raise ModelUnavailable(f"nạp VoxCPM2 thất bại: {type(exc).__name__}: {exc}") from exc
     log.info("voxcpm.load.done")
     return _model
@@ -140,7 +140,7 @@ class VoxCpmSynthesizer:
                 # thường là dấu hiệu video máy làm (F2.3). Kịch bản dài quá thì
                 # viết ngắn lại, việc đó do use case quyết định.
             )
-        except Exception as exc:  # noqa: BLE001
+        except Exception as exc:
             raise TtsFailed(f"VoxCPM2 sinh giọng thất bại: {type(exc).__name__}: {exc}") from exc
 
         _write_wav(wav, dest, sample_rate=getattr(model, "sample_rate", 16000))
@@ -158,7 +158,7 @@ def _write_wav(samples: Any, dest: Path, *, sample_rate: int) -> None:
         flat = [float(x) for x in samples]
 
     pcm = array.array(
-        "h", (max(-32768, min(32767, int(round(x * 32767)))) for x in flat)
+        "h", (max(-32768, min(32767, round(x * 32767))) for x in flat)
     )
     with wave.open(str(dest), "wb") as w:
         w.setnchannels(1)
